perf: BPF — shrink syscall latency to u32 (32→28 byte events)
Clamp syscall latency to u32 nanoseconds (~4.3s max), saving 4 bytes
per syscall event through the ring buffer. Parser widens back to u64.

+++ bpf/include/observoor.h
@@ -46,10 +46,10 @@ struct event_header {
     __u8  pad[6];
 } __attribute__((packed));
 
-// Syscall event (32 bytes total).
+// Syscall event (28 bytes total).
 struct syscall_event {
     struct event_header hdr;
-    __u64 latency_ns;
+    __u32 latency_ns;
 };
 
 // Disk I/O event (48 bytes total).
